requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dataclasses-json==0.6.7
datasets==2.21.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
evaluate==0.4.2
exceptiongroup==1.2.2
executing==2.0.1
faiss-gpu==1.7.2
fastapi==0.112.1
fastapi-cli==0.0.5
ffmpy==0.4.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-translate==3.16.0
googleapis-common-protos==1.65.0
gradio==4.42.0
gradio_client==1.3.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.4
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.38
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langcodes==3.4.0
langgraph==0.2.3
langgraph-checkpoint==1.0.9
langgraph-checkpoint-sqlite==1.0.3
langsmith==0.1.99
language_data==1.2.0
lightgbm==4.5.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.5.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.5.2
networkx==3.3
nltk==3.9.1
nmslib==2.0.6
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
onnxruntime==1.19.0
openai==1.43.0
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pillow==10.4.0
portalocker==2.10.1
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.5
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyjnius==1.6.1
pymongo==4.8.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyserini==0.37.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.20.0
rsa==4.9
ruff==0.6.1
sacrebleu==2.4.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.2
streamlit==1.37.1
streamlit-chat==0.1.1
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
triton==3.0.0
typer==0.12.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
wasabi==1.1.3
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
